requirements.txt
mcp-use
langchain-anthropic
python-dotenv